membership card member manage

+++ homework/time-tennis/src/main/java/com/lazyman/timetennis/menbership/MembershipCardManageController.java
@@ -1,9 +1,14 @@
 package com.lazyman.timetennis.menbership;
 
-import com.lazyman.timetennis.privilege.PrivilegeTool;
+import com.lazyman.timetennis.Constant;
+import com.lazyman.timetennis.arena.ArenaPrivilege;
 import com.lazyman.timetennis.user.User;
+import com.lazyman.timetennis.user.UserMapper;
+import com.lazyman.timetennis.wp.WePayService;
 import com.wisesupport.commons.exceptions.BusinessException;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Validate;
+import org.hibernate.validator.constraints.Length;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,12 +20,24 @@ import java.util.List;
 public class MembershipCardManageController {
     private MembershipCardDao mcDao;
 
-    public MembershipCardManageController(MembershipCardDao mcDao) {
+    private MembershipCardBillDao billDao;
+
+    private ArenaPrivilege privilege;
+
+    private UserMapper userMapper;
+
+    private MembershipCardService cardService;
+
+    public MembershipCardManageController(MembershipCardDao mcDao, MembershipCardBillDao billDao, ArenaPrivilege privilege, UserMapper userMapper, MembershipCardService cardService) {
         this.mcDao = mcDao;
+        this.billDao = billDao;
+        this.privilege = privilege;
+        this.userMapper = userMapper;
+        this.cardService = cardService;
     }
 
     @GetMapping("/metas")
-    public List<MembershipCardMeta> metas(User user, int arenaId) {
+    public List<MembershipCardMeta> metas(int arenaId) {
         return mcDao.byArenaId(arenaId);
     }
 
@@ -28,12 +45,12 @@ public class MembershipCardManageController {
     @Transactional
     public void deleteMeta(User user, @PathVariable int id) {
         MembershipCardMeta meta = mcDao.loadMeta(id);
-        Validate.notNull(meta);
+
+        privilege.requireAdministrator(user.getOpenId(), meta.getArena().getId());
 
         if (mcDao.hasMember(id)) {
             throw new BusinessException("该会员卡已经有会员,无法删除!");
         }
-        //todo 判断是不是场馆管的理员
         mcDao.deleteMeta(id, meta.getArena().getId());
     }
 
@@ -46,9 +63,9 @@ public class MembershipCardManageController {
                            @RequestParam int discount,
                            @RequestParam int price,
                            @RequestParam int extendMonth) {
-        //todo 权限
+
         MembershipCardMeta meta = mcDao.loadMeta(id);
-        Validate.notNull(meta);
+        privilege.requireAdministrator(user.getOpenId(), meta.getArena().getId());
         Validate.isTrue(mcDao.updateMeta(id, meta.getArena().getId(), name, initialBalance, discount, price, extendMonth) == 1);
     }
 
@@ -61,6 +78,7 @@ public class MembershipCardManageController {
                         @RequestParam int discount,
                         @RequestParam int price,
                         @RequestParam int extendMonth) {
+        privilege.requireAdministrator(user.getOpenId(), arenaId);
         mcDao.createMeta(arenaId, name, initialBalance, discount, price, extendMonth);
     }
 
@@ -74,10 +92,10 @@ public class MembershipCardManageController {
     public void status(User user,
                        @RequestParam boolean online,
                        @RequestParam int id) {
-
-        //todo 权限
         MembershipCardMeta meta = mcDao.loadMeta(id);
-        Validate.notNull(meta);
+
+        privilege.requireAdministrator(user.getOpenId(), meta.getArena().getId());
+
         if (online) {
             mcDao.metaOnline(meta.getArena().getId(), id);
         } else {
@@ -89,8 +107,42 @@ public class MembershipCardManageController {
     public List<MembershipCard> members(User user, int metaId,
                                         @RequestParam(required = false) String name,
                                         @RequestParam(required = false) String openId) {
-        //todo 权限
-        PrivilegeTool.assertHasArenaManagerRole(user);
-        return mcDao.members(metaId, name, openId);
+        MembershipCardMeta meta = mcDao.loadMeta(metaId);
+        privilege.requireAdministrator(user.getOpenId(), meta.getArena().getId());
+
+        if (StringUtils.isNumeric(name)) {
+            return mcDao.members(metaId, null, name, openId);
+        } else {
+            return mcDao.members(metaId, name, null, openId);
+        }
+    }
+
+    @PostMapping("/add_member")
+    @Transactional
+    public void addMember(User user,
+                          @RequestParam @NotEmpty String openId,
+                          @RequestParam int metaId) {
+        MembershipCardMeta meta = mcDao.loadMeta(metaId);
+        privilege.requireAdministrator(user.getOpenId(), meta.getArena().getId());
+
+        Validate.notNull(userMapper.selectByPrimaryKey(openId), "用户ID[%s]不存在", openId);
+        if (mcDao.hasMeta(openId, meta.getId())) {
+            throw new BusinessException("用户已经购买了该会员卡");
+        }
+        cardService.create(meta, openId);
+    }
+
+    @PutMapping("/recharge")
+    @Transactional
+    public void recharge(User user,
+                         @RequestParam @Length(min = 10, max = 10) String code,
+                         @RequestParam int fee) {
+        MembershipCard card = mcDao.loadCard(code);
+        MembershipCardMeta meta = mcDao.loadMeta(card.getMeta().getId());
+        privilege.requireAccountant(user.getOpenId(), meta.getArena().getId());
+
+        int balance = mcDao.recharge(code, fee);
+        String tradeNo = WePayService.creatTradeNo(Constant.PRODUCT_RECHARGE);
+        billDao.add(tradeNo, user.getOpenId(), code, Constant.PRODUCT_RECHARGE, fee, balance);
     }
 }
